Added logic for favorites, and changed the UI with displayCards to show a favorite button.

+++ utils/functions.php
@@ -178,11 +178,15 @@ function displayCards($recipes)
         <p class="author">Author: <?= $recipes[$i]['author'] ?></p>
       </a>
 
-      <?php if (allowedToEdit($recipes[$i]['author'])) { ?>
+      <?php
+      if (isset($_SESSION['signedIn'])) { ?>
         <div class="d-flex btns" id="btn-box-<?= $recipes[$i]['id'] ?>">
-          <a href="../entity/delete.php?recipe_id=<?= $recipes[$i]['id'] ?>" class="btn btn-danger btn-delete">Delete</a>
-          <a href="../entity/edit.php?recipe_id=<?= $recipes[$i]['id'] ?>" class="btn btn-secondary update-btn">Edit</a>
-        </div>
+        <?php if (allowedToEdit($recipes[$i]['author'])) { ?>
+            <a href="../entity/delete.php?recipe_id=<?= $recipes[$i]['id'] ?>" class="btn btn-danger btn-delete">Delete</a>
+            <a href="../entity/edit.php?recipe_id=<?= $recipes[$i]['id'] ?>" class="btn btn-secondary update-btn">Edit</a>
+        <?php } ?>
+        <a href="../favorites/add-favorite.php?recipe_id=<?= $recipes[$i]['id'] ?>" class="btn btn-secondary update-btn">Favorite</a>
+      </div>
       <?php } ?>
     </div>
     <?php
